Function cannot be called before it is defined
Currently we walk through all functions and constructors, and emit
the function declaration and definition at the same time. This means
that if a function tries to call a function which has not been emitted
yet, it will go bang.

So, first emit function declarations and then the definitions.

Found by with a new test case for revert().

Signed-off-by: Sean Young <sean@mess.org>