Implement constant folding optimization pass
This adds an optimization pass for constant folding. llvm does do constant
folding, but this is on a lower level, and cannot constant fold everything,
e.g. 256 bit multiply or cryptographic hash functions. For example:

	bytes32 hash = keccak256('f(int32)');

Is now constant folded and computed at compile time.

This also means we can detect problems where before we could not.

	function arg(int32 a) public returns (int32) {
		int32 x = 0;

		return a / x;
	}

This now generates a "division by zero" error, where before this was
a runtime error.

Signed-off-by: Sean Young <sean@mess.org>