Try to resolve expression to the expected type
As a simple example:

	uint16 x = 100 * 12;

The expression "100 * 12" should be resolved with the expectation that the
result will be a uint16. This means that the literals and the multiplication
are all of type uint16.

Before this commit, literals will always the smallest type that could
represent their value, and the complete expression was cast to the correct
type. This means that the expression used to be two uint8 literals, with a
uint8 multiply and the result cast to uint16. This is not correct.

This also solves the issue of having to cast the first element of an
array literal. If from the context it is clear what the elements of an
array should be, then all elements are cast to that type.

Lastly we can now correctly handle the hex literal value in bytesN
context, for example:

	bytes3 x = 0x001100;

This syntax is only permitted in Solidity if the number of digits matches
the length of the type. Now we know what the type is at resolve time this
is now processed correctly.

Signed-off-by: Sean Young <sean@mess.org>