Reference types can be null (string, bytes, arrays, structs)
When a struct or array is allocated, this is done with a single
malloc. For example an array of 4 structs is allocated in with a
single malloc, and the pointer to the structs are all set to null.

When dereferencing the array elements, now the structs are
automagically allocated.

In addition to that, the abi encoder can come across null pointers
for structs/arrays/bytes which haven't been allocated yet.

Signed-off-by: Sean Young <sean@mess.org>