Solana contract storage array length
The codegen for storage arrays must be removed from sema, as it does
not belong there and it should be target dependant; Solana requires
different handling. A single Expression::StorageArrayLength is used
for storage bytes, bytesN, and all arrays.

This also removes the need of for Expression::StorageBytesLength
which is now removed.

Signed-off-by: Sean Young <sean@mess.org>