Enfore the payable attribute on Solana
Functions or constructors not marked payable should abort if value is
transferred to them.

Note that on EVM, the check for value is done very early, to avoid
wasting gas on any work before aborting. Since on Solana this error
is fatal to the transaction anyway, this check can be done later.

Signed-off-by: Sean Young <sean@mess.org>