uint[1.111111E1111111111111] causes compiler to go into infinite loop
The lexer tries to calculate 10^1111111111111 which will be a very large
number. This will take a long time.

The lexer should not try to calculate numbers from strings, as this
might fail and we do not want the lexer to give up, as no parsing and
semantic analysis will be done.

Signed-off-by: Sean Young <sean@mess.org>