Refactor SCALE encoder into codegen (#1128)
This PR makes the AbiEncoding trait more granular. This allowed to move most of the logic into default implementations, leaving only a few Types that needed to be handled differently in SCALE and Borsh. This means the Borsh and SCALE encoders can share most of their code.

The decoding side is untouched by this PR as it can be done independent.