Expand list of solang code owners
When creating a pull request, the users in this files are used to populate
the reviewers. If it lists @hyperledger/solang-committers then we can't see
who those users are and as soon as one of the members of
@hyperledger/solang-committers leaves a review, then the pull request is
marked as approved while we really want *all* the maintainers to approve.

Signed-off-by: Sean Young <sean@mess.org>