Add --contract and --output-meta CLI flags (#1164)
Add cli option --output-meta directory

This sets the directory for the Solana idl file, or Substrate contract file.

Add --contract command line option to specify which contracts we want

The use cases is in the anchor environment. We have a solidity file,
and a name of contract. If that contract is not present in the source
code, we want compilation to fail.

Signed-off-by: Sean Young <sean@mess.org>